Migrate CI to GitHub Actions
Travis CI has been dead for two years now ... thank you for your service
Travis, you transformed the world of open source for the better.

While at it, also add a job to test Wasm, since that is supported too
now, but there we can't run the tests easily, so skip that.