Replace winapi dependency with windows-sys
Despite the former working just fine, it has not been updated in a few
years, and Microsoft published the windows-sys crate. The replacement
feels like pointless churn to me, but if the entire ecosystem is really
making the switch, then who am I to be the node in the graph that is
keeping the winapi dependency alive?