Bump GitHub Actions checkout action to 4.2
The 3.5 one is causing nodejs deprecation warnings. Because obviously
you need nodejs to check out a Git repository, and you need your users
to bump the number every 6 months despite there having been no
functional changes for the past several years (or even since launch for
my simple case).

Fortunately, this is now generated with RCL so I only have to bump the
number in *one* place instead of having to bump it in 12 places.