Bump version to 5.0.0
I consider it a breaking change because it bumps the MSRV. Also, there
is no point in updating unless you want to get rid of the "winapi" crate
in your dependency tree, but probably for many users this update will do
the opposite: you have the "winapi" dependency anyway, and this now adds
an additional dependency on "windows-sys" that may not have been there
before. So probably it's best to make a conscious decision about this
to bump the version, rather than getting a larger lockfile as an
unwelcome surprise when updating some packages.